+++ conftest.py
@@ -1,0 +1,1 @@
+pytest_plugins = ["model_runner.validator._tests.validator_test_fixtures"]
